requirements.txt
streamlit
pillow
numpy
openai==0.28
python-dotenv
pandas